# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A PLATFORM OPERATIONS API in golang using Gin framework",
        "title": "PLATFORM OPERATIONS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/process": {
            "post": {
                "description": "Create a Process and saves it in postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new Process",
                "parameters": [
                    {
                        "description": "Create Process",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateProcessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerProcess"
                        }
                    }
                }
            }
        },
        "/process/{id}": {
            "delete": {
                "description": "Get a Process from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "delete a Process with a specified id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Process ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update a Process from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "update a Process with a specified id",
                "parameters": [
                    {
                        "description": "update Process",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateProcessRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Process ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/processes": {
            "get": {
                "description": "List Processs according to the page_id and page_size",
                "produces": [
                    "application/json"
                ],
                "summary": "List all Processs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "api.CreateProcessRequest": {
            "type": "object",
            "properties": {
                "argo_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "template_id": {
                    "type": "integer"
                }
            }
        },
...
        "api.SwaggerProcess": {
            "type": "object",
            "properties": {
                "argo_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "template_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdateProcessRequest": {
            "type": "object",
            "properties": {
                "argo_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "template_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}